refactor unittest framework
as a result, there's another SUITE that needs to run
